Formatted the input forms and their labels using a grid.
Formatted the input forms and their labels using a grid. Will probably do something similar for the table if I have time.

Currently trying to make a custom CSS file work. Once it's complete I'll send it to you and we can adjust the path so the html can find it.